+++ blocks/src/blocks/components/count-card/budget/budget.ts
@@ -1,0 +1,4 @@
+import { IconRegistry } from '@tylertech/forge/icon';
+import { tylIconAccountBalance, tylIconAttachMoney, tylIconEdit, tylIconStar, tylIconTrendingDown } from '@tylertech/tyler-icons';
+
+IconRegistry.define([tylIconAccountBalance, tylIconAttachMoney, tylIconEdit, tylIconStar, tylIconTrendingDown]);