+++ blocks/src/blocks/components/structured-card/with-form/with-form.ts
@@ -1,0 +1,4 @@
+import { IconRegistry } from '@tylertech/forge/icon';
+import { tylIconConstruction, tylIconDelete, tylIconSave } from '@tylertech/tyler-icons';
+
+IconRegistry.define([tylIconConstruction, tylIconDelete, tylIconSave]);