+++ packages/forge/custom-elements-manifest.config.js
@@ -44,6 +44,9 @@ export default {
         state: {
           mappedName: 'states',
           isArray: true
+        },
+        meta: {
+          isArray: true
         }
       }
     })