+++ packages/forge/src/lib/quantity-field/quantity-field-constants.ts
@@ -1,0 +1,8 @@
+import { COMPONENT_NAME_PREFIX } from '../constants.js';
+
+const elementName: keyof HTMLElementTagNameMap = `${COMPONENT_NAME_PREFIX}quantity-field`;
+
+/** @deprecated - These are internal constants that will be removed/moved in the future. Please avoid using them. */
+export const QUANTITY_FIELD_CONSTANTS = {
+  elementName
+};